requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.1.4
cssselect==1.0.3
decorator==4.2.1
diff-match-patch==20121119
Django==1.9
django-crispy-forms==1.7.0
django-formtools==2.1
django-import-export==1.0.0
django-reversion==2.0.13
djangorestframework==3.7.7
dnspython==1.15.0
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipython==6.2.1
ipython-genutils==0.2.0
jdcal==1.3
jedi==0.11.1
lxml==4.1.1
odfpy==1.3.6
openpyxl==2.5.0
parsel==1.4.0
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
pyOpenSSL==17.5.0
pytz==2018.3
PyYAML==3.12
queuelib==1.4.2
requests==2.18.4
Scrapy==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
tablib==0.12.1
traitlets==4.3.2
Twisted==17.9.0
unicodecsv==0.14.1
urllib3==1.22
w3lib==1.19.0
wcwidth==0.1.7
xadmin==0.6.1
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.4.3
